backend/requirements.txt
flask
opencv-python
flask-cors
Pillow
ultralytics